fix(accordion): bloom accordion problems resolve
