+++ lib/MBMigration/Builder/Layout/Theme/Bloom/Assets/src/Accordion/index.ts
@@ -1,0 +1,1 @@
+export { getAccordion as run } from "elements/src/Accordion";
