+++ packages/elements/src/Tabs/index.ts
@@ -1,4 +1,4 @@
-import { getModel } from "./utils/getModel";
+import { getModel } from "../utils/getModel";
 import { Entry, Output } from "elements/src/types/type";
 import { createData } from "elements/src/utils/getData";
 import { parseColorString } from "utils/src/color/parseColorString";
